backend/pyproject.toml
[tool.poetry]
name = "jampada"
version = "0.1.0"
description = ""
authors = ["Surawut"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.91.0"
uvicorn = "^0.20.0"
mongoengine = "^0.26.0"
passlib = "^1.7.4"
loguru = "^0.6.0"
python-jose = "^3.3.0"
pydantic =  { version = "^1.9", extras = ["dotenv"] }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
